plugins/geobaserow/backend/requirements/base.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile --output-file=base.txt base.in
#
-e /baserow/backend
    # via -r base.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools

djangorestframework-gis==1.0
